requirements.txt
django
djangorestframework
django-json-widget
python-dotenv
openai
anthropic
google-generativeai
transformers
qdrant-client
sentence-transformers
tiktoken
pdfplumber
python-docx
open-clip-torch
pillow
pandas
psycopg2-binary
torch
